Add files via upload
Simply solving the Laplacian with Backslash - nothing fancy and no changing Dx or Dy step sizes, this simply has them as equal, though feel free to change them and play around. Both the numerical and the analytic solution are plotted (in that order), then the relative (non-absolute) error too, will shortly be looking to play around with a GIF to show how error changes as Dx and Dy decrease, but only after CG implemented and analysed.