Add files via upload
This is with a naive (non-averaging) restriction operator - note that when it is run, we still miss out on the exact solution by a fraction - possibly down to the naive implementation of the restrictor. Any thoughts, do let me know! Note, you can bump this up to i = 8 and it can handle it pretty well!